Drop-through @parameter fix.
a=#debug(obj="@obj");
#a=a;

b=#a(obj=#object().get(z="")) //Forcing a null.
/b = b;
c=b.call(obj="foo");
/c = c;

What do you expect the result is for /b and /c?

If you said that /b should be null and /c should be null, you are probably rational.

Currently, /c will debug and return 'foo' because #call (and function invocation) replaces an @param IFF it is not null.

This is very bad behavior, and I have now fixed it, but this is one of those 'version-increment worthy' changes.